fix Task router HTTP request methods

+++ routes/task.js
@@ -11,9 +11,8 @@ taskRouter.use(bodyParser.json());
 taskRouter
     .route('/')
     .get(authenticate.verifyUser, (req, res, next) => {
-        //Task.find()   //returns promise
-        Task.find({"author": req.user._id})   //returns tasks written by the current user
-            .populate('author')
+        Task.find({ user: req.user._id })
+            .populate('user')
                 .then(tasks => {
                     res.statusCode = 200;
                     res.setHeader('Content-Type', 'application/json');
@@ -22,7 +21,7 @@ taskRouter
                 .catch(err => next(err));   //pass err off to the overall error handler for this express application
     })
     .post(authenticate.verifyUser, (req, res, next) => {
-        req.body.author = req.user._id;
+        req.body.user = req.user._id;
         Task.create(req.body)   //returns promise (mongoose checks if data fits schema)
             .then(task => {
                 console.log('Task Created: ', task);
@@ -37,7 +36,7 @@ taskRouter
         res.end('PUT operation not supported on /tasks');
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Task.deleteMany()
+        Task.deleteMany({ user: req.user._id })
             .then(response => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -50,39 +49,74 @@ taskRouter
     .route('/:taskId')
     .get(authenticate.verifyUser, (req, res, next) => {
         Task.findById(req.params.taskId)
-            .populate('author')
-                .then(task => {
-                    res.statusCode = 200;
-                    res.setHeader('Content-Type', 'application/json');
-                    res.json(task);
-                })
-                .catch(err => next(err));
+            .then(task => {
+                if (task) {
+                    if (task.user._id.equals(req.user._id)) {
+                        res.statusCode = 200;
+                        res.setHeader('Content-Type', 'application/json');
+                        res.json(task);
+                    } else {
+                        res.statusCode = 403;
+                        res.end('This is not your task!');
+                    }
+                } else {
+                    res.statusCode = 403;
+                    res.end('Task does not exist!');
+                }
+            })
+            .catch(err => next(err));
     })
     .post(authenticate.verifyUser, (req, res) => {
         res.statusCode = 403;
         res.end(`POST operation not supported on /tasks/${req.params.taskId}`);
     })
     .put(authenticate.verifyUser, (req, res, next) => {
-        Task.findByIdAndUpdate(
-            req.params.taskId, 
-            {
-                $set: req.body
-            }, 
-            { new: true }
-        )
+        Task.findById(req.params.taskId)
             .then(task => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(task);
+                if (task) {
+                    if (task.user._id.equals(req.user._id)) {
+                        Task.findByIdAndUpdate(
+                            req.params.taskId, 
+                            { $set: req.body }, 
+                            { new: true }
+                        )
+                            .then(task => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(task);
+                            })
+                            .catch(err => next(err));
+                    } else {
+                        res.statusCode = 403;
+                        res.end('This is not your task!');
+                    }
+                } else {
+                    res.statusCode = 403;
+                    res.end('Task does not exist!');
+                }
             })
             .catch(err => next(err));
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Task.findByIdAndDelete(req.params.taskId)
-            .then(response => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(response);
+        Task.findById(req.params.taskId)
+            .then(task => {
+                if (task) {
+                    if (task.user._id.equals(req.user._id)) {
+                        Task.findByIdAndDelete(req.params.taskId)
+                            .then(response => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(response);
+                            })
+                            .catch(err => next(err));
+                    } else {
+                        res.statusCode = 403;
+                        res.end('This is not your task!');
+                    }
+                } else {
+                    res.statusCode = 403;
+                    res.end('Task does not exist!');
+                }
             })
             .catch(err => next(err));
     });
